refactor(orchestrator): fromAppSettings and CDN config belong on base class
fromAppSettings is product logic, not debug instrumentation. Moved to
LodArtworkOrchestrator with a protected buildAppSettingsConfig() helper so
the debug subclass can override fromAppSettings() to return its own type
without duplicating the AppSettings/CDN-ceiling logic.

Also removes the vestigial maxGames field from LodArtworkOrchestratorDebug
(it was stored but never read — maxTextures on the base class already
sizes the texture array and instance buffer).
